requirements.txt
aiohttp==3.9.3
python-dotenv==1.0.1
websockets==12.0
discord-webhook==1.3.0 